RUN-34 Refactor exercises structure
