+++ bases/rsptx/assignment_server_api/assignment_builder/src/components/routes/AssignmentBuilder/components/exercises/components/CreateExercise/components/ClickableAreaExercise/ClickableAreaExercise.tsx
@@ -1,4 +1,4 @@
-import { ExerciseComponentProps } from "../types/ExerciseTypes";
+import { ExerciseComponentProps } from "../../types/ExerciseTypes";
 
 export const ClickableAreaExercise = ({}: ExerciseComponentProps) => {
   return <></>;
