+++ bases/rsptx/assignment_server_api/assignment_builder/src/components/routes/AssignmentBuilder/components/exercises/components/CreateExercise/components/ShortAnswerExercise/ShortAnswerExercise.tsx
@@ -1,0 +1,6 @@
+import { ExerciseComponentProps } from "@components/routes/AssignmentBuilder/components/exercises/components/CreateExercise/types/ExerciseTypes";
+import { FC } from "react";
+
+export const ShortAnswerExercise: FC<ExerciseComponentProps> = () => {
+  return <></>;
+};
