+++ bases/rsptx/assignment_server_api/assignment_builder/src/components/routes/AssignmentBuilder/components/exercises/components/CreateExercise/components/index.ts
@@ -1,10 +1,10 @@
 export { ExerciseFactory } from "./ExerciseFactory";
 export { BaseExerciseForm } from "./BaseExerciseForm";
-export { ParsonsExercise } from "./ParsonsExercise";
-export { ActiveCodeExercise } from "./ActiveCodeExercise";
-export { FillInTheBlankExercise } from "./FillInTheBlankExercise";
-export { DragAndDropExercise } from "./DragAndDropExercise";
-export { ClickableAreaExercise } from "./ClickableAreaExercise";
+export { ParsonsExercise } from "./ParsonsExercise/ParsonsExercise";
+export { ActiveCodeExercise } from "./ActiveCodeExercise/ActiveCodeExercise";
+export { FillInTheBlankExercise } from "./FillInTheBlankExercise/FillInTheBlankExercise";
+export { DragAndDropExercise } from "./DragAndDropExercise/DragAndDropExercise";
+export { ClickableAreaExercise } from "./ClickableAreaExercise/ClickableAreaExercise";
 export * from "./PollExercise/PollExercise";
-export { ShortAnswerExercise } from "./ShortAnswerExercise";
+export { ShortAnswerExercise } from "./ShortAnswerExercise/ShortAnswerExercise";
 export * from "./MultiChoiceExercise";
